+++ src/workshop_3/controller/SolutionsAdminDelete.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import workshop_3.DAO.SolutionDAO;
+import workshop_3.misc.ValPar;
 import workshop_3.model.Solution;
 
 @WebServlet("/deletesolution")
@@ -21,18 +22,12 @@ public class SolutionsAdminDelete extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		String idParam = request.getParameter("id").toString();
-		int solutionId = 0;
-		if(idParam!=null & idParam!="") {
-			try {
-				solutionId=Integer.parseInt(idParam);
-			}catch(NumberFormatException e) {
-				System.out.println("Incorrect solution Id!");
-				e.printStackTrace();
-			}
+		long solutionId = ValPar.longVar(idParam, "Incorrect solution Id!");
+		if(solutionId >= 0) {
+			SolutionDAO solutionDAO=new SolutionDAO();
+			Solution solution=SolutionDAO.loadSolutionById(solutionId);
+			solutionDAO.deleteSolution(solution);
 		}
-		SolutionDAO solutionDAO=new SolutionDAO();
-		Solution solution=SolutionDAO.loadSolutionById(solutionId);
-		solutionDAO.deleteSolution(solution);
 		getServletContext().getRequestDispatcher("/solutionsadminpanel").forward(request, response);
 	}
